styling overhaul (variables, dark mode, rems, …)
These are all quite subjective but I wanted to improve a number of
things with the site that I think improve the experience

- move to CSS variable (with fallbacks)
- support dark themes via media query
    - code blocks with a custom theme
    - thinner weight font
    - more saturation in colors
- base most unit on `rem` to support users size preferences
- limit some elements to an optimal character limit per line
- padding around edges for a better 100% (e.g. mobile) experience
- increase line leading
- use `system-ui` OS font (falling back to `sans-serif`)
- nav links give feedback with an underline on `:focus` & `:hover`
- header+footer have a gradient that matches the accent color scheme
- remove justified text because rivers are awkward with code blocks

The biggest todo to ‘finish’ this would be to replace the header image
with an SVG instead of PNG. Currently the black does not show up. This
SVG should have a media query inside of it to reverse the color of the
text on `prefers-color-sheme: dark`. This said, I think it’s overal an
improvement to support dark themes now and cross that later.

If you want to test the dark theme, in Firefox there’s a side panel
under the “Inspector” tab where there i as way to filter styles; this
same bar also has sun/moon icons that can toggle the color scheme
preference.