Add contribute.json
https://www.contributejson.org/

This spec is used in a lot of open source projects to signal to users
how to contribute to a given project. It originated at Mozilla but I
have seen it in the wild on a number of open-source projects. Since it is
defined as a schema, this file could in the future serve as a source of
truth for the `/contributing` and `/support` pages as well as others, but I
think just having it exist could be useful to some.

The original source for this file however was in Dhall (because of
course it was):

```dhall
let Contribute = https://git.sr.ht/~toastal/dhall-contribute-json/blob/trunk/Contribute.dhall

in  Contribute::{
    , name = "Soupault"
    , description =
        "Soupault is a tool that helps you create and manage static websites."
    , repository = Contribute.Repository::{
      , type = Some "git"
      , url = "https://github.com/dmbaturin/soupault"
      , license = "MIT"
      , clone = Some "git@github.com:dmbaturin/soupault.git"
      }
    , bugs = Some Contribute.Bugs::{
      , list = Some "https://github.com/dmbaturin/soupault/issues"
      , report = Some "https://github.com/dmbaturin/soupault/issues/new"
      }
    , participate = Some Contribute.Participate::{
      , home = Some "https://soupault.app/support/"
      , docs = "https://soupault.app/reference-manual/"
      , mailing-list = Some "https://lists.sr.ht/~dmbaturin/soupault"
      , irc = Some "https://libera.chat/#soupault"
      }
    , keywords = Some
      [ "static-site-generator", "web", "lua", "ocaml", "html-processor" ]
    }
```