SourceHut quick link revision
SourceHut supports Git, Mercurial, with flexibility to support more in
the future (such as approval for darcs if someone does the work) so it’s
not safe to assume ‘Git’. Additionally, “projects” do not have a VCS
subdomain (meant to be the home for mailing list, issues, & source)
which probably _should_ be the default.

Another error is assuming the leading `~` sigil. They seem pretty
adamant that a user’s ID includes this sigil (i.e. my user is `~toastal`
*not* `toastal`). I had heard in the future there was a plan to use a
different sigil to signify groups/teams—a feature currently missing from
the SourceHut forge. It’s better to make users add this.